@@ -9,29 +9,30 @@
     }
   },
   "dependencies": {
-    "@angular/common": "^5.2.10",
-    "@angular/compiler": "^5.2.10",
-    "@angular/core": "^5.2.10",
-    "@angular/forms": "^5.2.10",
-    "@angular/platform-browser": "^5.2.10",
-    "@angular/platform-browser-dynamic": "^5.2.10",
-    "@angular/router": "^5.2.10",
-    "@angular/service-worker": "^5.2.10",
-    "@ngrx/effects": "^5.2.0",
-    "@ngrx/store": "^5.2.0",
-    "json-midi-encoder": "^2.1.27",
-    "midi-json-parser": "^4.2.18",
-    "midi-json-parser-worker": "^4.0.19",
-    "rxjs": "^5.5.10",
-    "rxjs-broker": "^2.2.24",
-    "tslib": "^1.9.0",
+    "@angular/common": "7.0.4",
+    "@angular/compiler": "7.0.4",
+    "@angular/core": "7.0.4",
+    "@angular/forms": "7.0.4",
+    "@angular/platform-browser": "7.0.4",
+    "@angular/platform-browser-dynamic": "7.0.4",
+    "@angular/router": "7.0.4",
+    "@angular/service-worker": "7.0.4",
+    "@ngrx/effects": "^6.1.2",
+    "@ngrx/store": "^6.1.2",
+    "json-midi-encoder": "^3.0.5",
+    "midi-json-parser": "^6.0.5",
+    "midi-json-parser-worker": "^6.0.3",
+    "rxjs": "^6.3.3",
+    "rxjs-broker": "^4.0.9",
+    "tslib": "^1.9.3",
     "zone.js": "^0.8.26"
   },
   "description": "This is the client for analog4all.",
   "devDependencies": {
-    "@angular/cli": "~1.7.4",
-    "@angular/compiler-cli": "^5.2.10",
-    "@angular/language-service": "^5.2.10",
+    "@angular-devkit/build-angular": "^0.10.6",
+    "@angular/cli": "7.0.6",
+    "@angular/compiler-cli": "7.0.4",
+    "@angular/language-service": "7.0.4",
     "@commitlint/cli": "^7.2.1",
     "@commitlint/config-angular": "^7.1.2",
     "@hint/formatter-stylish": "^2.0.0",
@@ -62,16 +63,15 @@
     "@hint/hint-validate-set-cookie-header": "^2.0.0",
     "@hint/hint-x-content-type-options": "^2.0.0",
     "@hint/parser-manifest": "^2.0.0",
-    "@ngrx/store-devtools": "^5.2.0",
+    "@ngrx/store-devtools": "^6.1.2",
     "@types/deep-freeze-strict": "^1.1.0",
     "@types/jasmine": "~2.8.7",
     "@types/jasminewd2": "~2.0.2",
-    "@types/node": "~6.0.105",
+    "@types/node": "~8.9.5",
     "axe-core": "^3.1.2",
     "bundle-buddy": "^0.2.1",
     "commitizen": "^3.0.4",
     "content-security-policy-builder": "^2.0.0",
-    "core-js": "^2.5.5",
     "cz-conventional-changelog": "^2.1.0",
     "deep-freeze-strict": "^1.1.1",
     "eslint": "^5.9.0",
@@ -92,24 +92,24 @@
     "gruntify-eslint": "^5.0.0",
     "hint": "^4.0.3",
     "husky": "^1.1.4",
-    "jasmine-core": "~2.8.0",
+    "jasmine-core": "~2.99.1",
     "jasmine-spec-reporter": "~4.2.1",
-    "karma": "~2.0.2",
+    "karma": "~3.0.0",
     "karma-chrome-launcher": "~2.2.0",
-    "karma-coverage-istanbul-reporter": "^1.4.2",
+    "karma-coverage-istanbul-reporter": "~2.0.4",
     "karma-jasmine": "~1.1.2",
     "karma-jasmine-html-reporter": "^0.2.2",
     "karma-sauce-launcher": "^1.2.0",
     "load-grunt-config": "^0.19.2",
     "ngrx-store-freeze": "^0.2.2",
-    "protractor": "~5.1.2",
+    "protractor": "~5.4.1",
     "stylelint": "^9.8.0",
     "stylelint-config-holy-grail": "^3.0.0",
-    "ts-node": "~4.1.0",
-    "tsconfig-holy-grail": "^4.0.10",
-    "tslint": "~5.10.0",
-    "tslint-config-holy-grail": "^25.0.0",
-    "typescript": "~2.6.2",
+    "ts-node": "~7.0.1",
+    "tsconfig-holy-grail": "^7.0.12",
+    "tslint": "~5.11.0",
+    "tslint-config-holy-grail": "^36.1.2",
+    "typescript": "~3.1.6",
     "webpack-bundle-analyzer": "^3.0.3",
     "webpack-stats-duplicates": "^2.1.0"
   },
