@@ -21,17 +21,17 @@ module.exports = {
         'sh:build',
         'replace:bundle',
         'copy:404',
-        'htmlmin',
-        'replace:inline',
+        'replace:runtime',
         'replace:chunks',
         'replace:csp-production',
-        'clean:inline',
+        'clean:runtime',
         'copy:scripts',
         'replace:scripts',
         'clean:scripts',
         'copy:styles',
         'replace:styles',
         'clean:styles',
+        'htmlmin',
         'replace:manifest',
         'gh-pages:deploy',
         'smoke'
