@@ -1,10 +1,12 @@
+const { defaultProject, projects } = require('../../angular.json');
+
 module.exports = {
     chrome: {
         options: {
             browser: 'chrome'
         },
         urls: [
-            'http://localhost:6699'
+            `http://localhost:${ projects[defaultProject].targets.serve.options.port }`
         ]
     }
 };
