@@ -4,7 +4,7 @@ const pkg = require('../../package.json');
 module.exports = {
     deploy: {
         options: {
-            base: 'build/',
+            base: 'build/analog4all-client',
             message: `build page for version ${ env.TRAVIS_TAG }`,
             repo: pkg.repository.url.replace(/:\/\//, `://${ env.GIT_HUB_ACCESS_TOKEN }@`),
             silent: true,
