@@ -21,7 +21,7 @@ module.exports = (grunt) => {
         'bundle': {
             files: {
                 './': [
-                    'build/main.*.bundle.js'
+                    'build/analog4all-client/main.*.js'
                 ]
             },
             options: {
@@ -34,27 +34,27 @@ module.exports = (grunt) => {
         'chunks': {
             files: {
                 './': [
-                    'build/index.html'
+                    'build/analog4all-client/index.html'
                 ]
             },
             options: {
                 patterns: [ {
-                    match: /""\+e\+"\."\+{([0-9]+:"[a-f0-9]{20}",?)+}/g,
-                    replacement: (match) => match.replace(/""\+e\+"/g, '"scripts/"+e+"')
+                    match: /""\+\({}\[e\]\|\|e\)\+"\."\+{([0-9]+:"[a-f0-9]{20}",?)+}/g,
+                    replacement: (match) => match.replace(/^""/g, '"scripts/"')
                 } ]
             }
         },
         'csp-production': {
             files: {
-                'build/index.html': [
-                    'build/index.html'
+                'build/analog4all-client/index.html': [
+                    'build/analog4all-client/index.html'
                 ]
             },
             options: {
                 patterns: [ {
                     match: /<meta\shttp-equiv="content-security-policy">/,
                     replacement: () => {
-                        const html = fs.readFileSync('build/index.html', 'utf-8');
+                        const html = fs.readFileSync('build/analog4all-client/index.html', 'utf-8');
                         const regex = /<script[^>]*?>([^<](.|[\n\r])*?)<\/script>/gm;
                         const scriptHashes = [];
 
@@ -84,57 +84,57 @@ module.exports = (grunt) => {
                 } ]
             }
         },
-        'inline': {
-            files: {
-                './': [
-                    'build/index.html'
-                ]
-            },
-            options: {
-                patterns: [ {
-                    match: /<script\stype="text\/javascript"\ssrc="(inline\.[a-z0-9]*\.bundle\.js)"\sintegrity="sha384-[a-zA-Z0-9+/]*=*"\scrossorigin="anonymous"><\/script>/g,
-                    replacement: (match, filename) => {
-                        return `<script type="text/javascript">${ fs.readFileSync(`build/${ filename }`) }</script>`;
-                    }
-                } ]
-            }
-        },
         'manifest': {
             files: {
                 './': [
-                    'build/ngsw.json'
+                    'build/analog4all-client/ngsw.json'
                 ]
             },
             options: {
                 patterns: [ {
                     match: /assets\/apple-touch-icon\.png/g,
-                    replacement: () => grunt.file.expand({ cwd: 'build' }, 'assets/*.apple-touch-icon.png')[0]
+                    replacement: () => grunt.file.expand({ cwd: 'build/analog4all-client' }, 'assets/*.apple-touch-icon.png')[0]
                 }, {
                     match: /assets\/favicon\.ico/g,
-                    replacement: () => grunt.file.expand({ cwd: 'build' }, 'assets/*.favicon.ico')[0]
+                    replacement: () => grunt.file.expand({ cwd: 'build/analog4all-client' }, 'assets/*.favicon.ico')[0]
                 }, {
-                    match: /\/([a-z0-9-]+\.[a-z0-9]*\.bundle\.css)"/g,
+                    match: /\/([a-z0-9-]+\.[a-z0-9]*\.css)"/g,
                     replacement: (_, filename) => `/styles/${ filename }"`
                 }, {
-                    match: /\/([a-z0-9-]+\.[a-z0-9]*\.(bundle|chunk)\.js)"/g,
+                    match: /\/([a-z0-9-]+\.[a-z0-9]*\.js)"/g,
                     replacement: (_, filename) => `/scripts/${ filename }"`
                 }, {
-                    match: /[\s]*"\/analog4all-client(\/scripts)?\/inline\.[a-z0-9]+.bundle.js",/g,
+                    match: /[\s]*"\/analog4all-client(\/scripts)?\/runtime\.[a-z0-9]+.js",/g,
                     replacement: ''
                 }, {
-                    match: /[\s]*"\/analog4all-client(\/scripts)?\/inline\.[a-z0-9]+.bundle.js":\s"[a-z0-9]+",/g,
+                    match: /[\s]*"\/analog4all-client(\/scripts)?\/runtime\.[a-z0-9]+.js":\s"[a-z0-9]+",/g,
                     replacement: ''
                 }, {
-                    // Replace the hash value inside of the hashTable for "/scripts/main.*.bundle.js" because it was modified before.
-                    match: /"\/analog4all-client(\/scripts\/main\.[a-z0-9]+.bundle.js)":\s"[a-z0-9]+"/g,
+                    // Replace the hash value inside of the hashTable for "/scripts/main.*.js" because it was modified before.
+                    match: /"\/analog4all-client(\/scripts\/main\.[a-z0-9]+.js)":\s"[a-z0-9]+"/g,
                     replacement: (_, filename) => {
-                        return `"/analog4all-client${ filename }": "${ computeHashOfFile(`build${ filename }`, 'sha1', 'hex') }"`;
+                        return `"/analog4all-client${ filename }": "${ computeHashOfFile(`build/analog4all-client${ filename }`, 'sha1', 'hex') }"`;
                     }
                 }, {
                     // Replace the hash value inside of the hashTable for "/index.html" because it was modified before.
                     match: /"\/analog4all-client\/index\.html":\s"[a-z0-9]+"/g,
                     replacement: () => {
-                        return `"/analog4all-client/index.html": "${ computeHashOfFile('build/index.html', 'sha1', 'hex') }"`;
+                        return `"/analog4all-client/index.html": "${ computeHashOfFile('build/analog4all-client/index.html', 'sha1', 'hex') }"`;
+                    }
+                } ]
+            }
+        },
+        'runtime': {
+            files: {
+                './': [
+                    'build/analog4all-client/index.html'
+                ]
+            },
+            options: {
+                patterns: [ {
+                    match: /<script\stype="text\/javascript"\ssrc="(runtime\.[a-z0-9]*\.js)"\sintegrity="sha384-[a-zA-Z0-9+/]*=*"\scrossorigin="anonymous"><\/script>/g,
+                    replacement: (match, filename) => {
+                        return `<script type="text/javascript">${ fs.readFileSync(`build/analog4all-client/${ filename }`) }</script>`;
                     }
                 } ]
             }
@@ -142,15 +142,15 @@ module.exports = (grunt) => {
         'scripts': {
             files: {
                 './': [
-                    'build/index.html'
+                    'build/analog4all-client/index.html'
                 ]
             },
             options: {
                 patterns: [ {
-                    match: /<script\stype="text\/javascript"\ssrc="([a-z-]*\.[a-z0-9]*\.bundle\.js)"\sintegrity="(sha384-[a-zA-Z0-9+/]*=*)"\scrossorigin="anonymous"><\/script>/g,
+                    match: /<script\stype="text\/javascript"\ssrc="([a-z-]*\.[a-z0-9]*\.js)"\sintegrity="(sha384-[a-zA-Z0-9+/]*=*)"\scrossorigin="anonymous"><\/script>/g,
                     replacement: (match, filename, initialHash) => {
-                        const updatedHash = (/main\.[a-z0-9]*\.bundle\.js/.test(filename)) ?
-                            `sha384-${ computeHashOfFile(`build/scripts/${ filename }`, 'sha384', 'base64') }` :
+                        const updatedHash = (/main\.[a-z0-9]*\.js/.test(filename)) ?
+                            `sha384-${ computeHashOfFile(`build/analog4all-client/scripts/${ filename }`, 'sha384', 'base64') }` :
                             initialHash;
 
                         return `<script type="text/javascript" src="scripts/${ filename }" integrity="${ updatedHash }" crossorigin="anonymous"></script>`;
@@ -161,12 +161,12 @@ module.exports = (grunt) => {
         'styles': {
             files: {
                 './': [
-                    'build/index.html'
+                    'build/analog4all-client/index.html'
                 ]
             },
             options: {
                 patterns: [ {
-                    match: /<link\shref="(styles\.[a-z0-9]*\.bundle\.css)"\srel="stylesheet"\sintegrity="(sha384-[a-zA-Z0-9+/]*=*)"\scrossorigin="anonymous"\/>/g,
+                    match: /<link\srel="stylesheet"\shref="(styles\.[a-z0-9]*\.css)"\sintegrity="(sha384-[a-zA-Z0-9+/]*=*)"\scrossorigin="anonymous">/g,
                     replacement: (match, filename, hash) => {
                         return `<link href="styles/${ filename }" rel="stylesheet" integrity="${ hash }" crossorigin="anonymous">`;
                     }
