@@ -1,12 +1,9 @@
+const { join } = require('path');
 const { env } = require('process');
 
 module.exports = (config) => {
 
     config.set({
-
-        angularCli: {
-            environment: 'dev'
-        },
 
         basePath: '../../',
 
@@ -15,38 +12,22 @@ module.exports = (config) => {
         },
 
         coverageIstanbulReporter: {
+            dir: join(__dirname, '../../coverage'),
             fixWebpackSourcePaths: true,
             reports: [ 'html', 'lcovonly' ]
         },
 
-        files: [
-            {
-                pattern: './config/karma/test.ts',
-                watched: false
-            }
-        ],
-
         frameworks: [
-            '@angular/cli',
+            '@angular-devkit/build-angular',
             'jasmine'
         ],
 
-        mime: {
-            'text/x-typescript': [ 'ts', 'tsx' ]
-        },
-
         plugins: [
-            '@angular/cli/plugins/karma',
+            '@angular-devkit/build-angular/plugins/karma',
             'karma-*'
         ],
 
-        preprocessors: {
-            './config/karma/test.ts': [ '@angular/cli' ]
-        },
-
-        reporters: config.angularCli && config.angularCli.codeCoverage
-            ? [ 'progress', 'coverage-istanbul' ]
-            : [ 'progress', 'kjhtml' ]
+        reporters: [ 'progress', 'kjhtml' ]
 
     });
 
@@ -74,8 +55,6 @@ module.exports = (config) => {
         });
     } else {
         config.set({
-
-            browserNoActivityTimeout: 20000,
 
             browsers: [
                 'ChromeHeadless'
