@@ -1,4 +1,4 @@
-import { Component, OnInit } from '@angular/core';
+import { Component } from '@angular/core';
 import { GeneratorsService, PeerConnectingService } from './shared';
 
 @Component({
@@ -6,16 +6,14 @@ import { GeneratorsService, PeerConnectingService } from './shared';
     styleUrls: [ './app.component.css' ],
     templateUrl: './app.component.html'
 })
-export class AppComponent implements OnInit {
+export class AppComponent {
 
     public isSupported: boolean;
 
     constructor (
         private _generatorsService: GeneratorsService,
         private _peerConnectingService: PeerConnectingService
-    ) { }
-
-    public ngOnInit () {
+    ) {
         this.isSupported = this._generatorsService.isSupported &&
             this._peerConnectingService.isSupported;
     }
