@@ -1,6 +1,6 @@
 import { Injectable } from '@angular/core';
+import { Observable } from 'rxjs';
 import { IMaskableSubject, IStringifyableJsonObject } from 'rxjs-broker';
-import { Observable } from 'rxjs/Observable';
 
 @Injectable()
 export class FileReceivingService {
